response entity and error handling

+++ src/main/resources/static/assets/scripts/products.js
@@ -25,19 +25,40 @@ class Product {
 saveProductForm.addEventListener('submit', event => {
     event.preventDefault();
     const productData = event.currentTarget.querySelectorAll('input');
+    const wrongProductInfoLabel = event.currentTarget.querySelector('#wrong-product-info');
 
-    // add validation to User
+    if (productData[0].value.length === 0 || productData[1].value.length === 0 || productData[2].value.length === 0) {
+        wrongProductInfoLabel.style.visibility = 'visible';
+        wrongProductInfoLabel.textContent = 'invalid product info';
+    } else {
+        wrongProductInfoLabel.style.visibility = 'hidden';
+        const product = new Product(productData[0].value, productData[1].value, productData[2].value);
+        httpHandler.sendHttpRequest('POST', httpHandler.urlPrefix + 'save-product', 'json', product)
+            .then(() => renderNewTableRow(productData[0].value, productData[1].value, productData[2].value));
+    }
 
-    const product = new Product(productData[0].value, productData[1].value, productData[2].value);
-    httpHandler.sendHttpRequest('POST', httpHandler.urlPrefix + 'save-product', 'json', product)
-        .then(() => renderNewTableRow(productData[0].value, productData[1].value, productData[2].value));
 })
+
+
+const isProductPresent = (productName) => {
+    if (productName.length === 0) return false;
+    return Boolean[table.querySelector(`#${productName}`)];
+}
 
 deleteProductForm.addEventListener('submit', event => {
     event.preventDefault();
     const productName = event.currentTarget.querySelector('input');
-    httpHandler.sendHttpRequest('POST', httpHandler.urlPrefix + 'delete-product?productName=' + productName.value, 'json')
-        .then(() => deleteTableRowByName(productName.value));
+    const wrongProductNameLabel = event.currentTarget.querySelector('#wrong-product-name');
+
+    if (!isProductPresent(productName.value)) {
+        wrongProductNameLabel.style.visibility = 'visible';
+        wrongProductNameLabel.textContent = 'invalid product name';
+    } else {
+        wrongProductNameLabel.style.visibility = 'hidden';
+        httpHandler.sendHttpRequest('DELETE', httpHandler.urlPrefix + 'delete-product?productName=' + productName.value, 'json')
+            .then(() => deleteTableRowByName(productName.value));
+    }
+
 })
 
 const renderNewTableRow = (productName, price, quantity) => {
